test (LineSectionAcceptanceTest) : static method 클래스 분리 리팩토링
