+++ src/test/java/nextstep/subway/line/acceptance/LineSectionAcceptanceTest.java
@@ -1,5 +1,6 @@
 package nextstep.subway.line.acceptance;
 
+import static nextstep.subway.line.acceptance.LineSectionStaticAcceptance.*;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.Arrays;
@@ -24,62 +25,12 @@ import nextstep.subway.station.dto.StationResponse;
 
 @DisplayName("지하철 구간 관련 기능")
 public class LineSectionAcceptanceTest extends AcceptanceTest {
+
 	private LineResponse 신분당선;
 	private StationResponse 강남역;
 	private StationResponse 양재역;
 	private StationResponse 정자역;
 	private StationResponse 광교역;
-
-	public static ExtractableResponse<Response> 지하철_노선에_지하철역_등록_요청(LineResponse line, StationResponse upStation,
-		StationResponse downStation, int distance) {
-		SectionRequest sectionRequest = new SectionRequest(upStation.getId(), downStation.getId(), distance);
-
-		return RestAssured
-			.given().log().all()
-			.contentType(MediaType.APPLICATION_JSON_VALUE)
-			.body(sectionRequest)
-			.when().post("/lines/{lineId}/sections", line.getId())
-			.then().log().all()
-			.extract();
-	}
-
-	public static void 지하철_노선에_지하철역_등록됨(ExtractableResponse<Response> response) {
-		assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-	}
-
-	public static void 지하철_노선에_지하철역_등록_실패됨(ExtractableResponse<Response> response) {
-		assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
-	}
-
-	public static void 지하철_노선에_지하철역_순서_정렬됨(ExtractableResponse<Response> response,
-		List<StationResponse> expectedStations) {
-		LineResponse line = response.as(LineResponse.class);
-		List<Long> stationIds = line.getStations().stream()
-			.map(it -> it.getId())
-			.collect(Collectors.toList());
-
-		List<Long> expectedStationIds = expectedStations.stream()
-			.map(it -> it.getId())
-			.collect(Collectors.toList());
-
-		assertThat(stationIds).containsExactlyElementsOf(expectedStationIds);
-	}
-
-	public static ExtractableResponse<Response> 지하철_노선에_지하철역_제외_요청(LineResponse line, StationResponse station) {
-		return RestAssured
-			.given().log().all()
-			.when().delete("/lines/{lineId}/sections?stationId={stationId}", line.getId(), station.getId())
-			.then().log().all()
-			.extract();
-	}
-
-	public static void 지하철_노선에_지하철역_제외됨(ExtractableResponse<Response> response) {
-		assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-	}
-
-	public static void 지하철_노선에_지하철역_제외_실패됨(ExtractableResponse<Response> response) {
-		assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
-	}
 
 	@BeforeEach
 	public void setUp() {
