+++ src/test/java/nextstep/subway/line/acceptance/LineSectionStaticAcceptance.java
@@ -1,0 +1,71 @@
+package nextstep.subway.line.acceptance;
+
+import static org.assertj.core.api.Assertions.*;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.line.dto.SectionRequest;
+import nextstep.subway.station.dto.StationResponse;
+
+public class LineSectionStaticAcceptance {
+
+	public static ExtractableResponse<Response> 지하철_노선에_지하철역_등록_요청(LineResponse line, StationResponse upStation,
+		StationResponse downStation, int distance) {
+		SectionRequest sectionRequest = new SectionRequest(upStation.getId(), downStation.getId(), distance);
+
+		return RestAssured
+			.given().log().all()
+			.contentType(MediaType.APPLICATION_JSON_VALUE)
+			.body(sectionRequest)
+			.when().post("/lines/{lineId}/sections", line.getId())
+			.then().log().all()
+			.extract();
+	}
+
+	public static void 지하철_노선에_지하철역_등록됨(ExtractableResponse<Response> response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+	}
+
+	public static void 지하철_노선에_지하철역_등록_실패됨(ExtractableResponse<Response> response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
+	}
+
+	public static void 지하철_노선에_지하철역_순서_정렬됨(ExtractableResponse<Response> response,
+		List<StationResponse> expectedStations) {
+		LineResponse line = response.as(LineResponse.class);
+		List<Long> stationIds = line.getStations().stream()
+			.map(it -> it.getId())
+			.collect(Collectors.toList());
+
+		List<Long> expectedStationIds = expectedStations.stream()
+			.map(it -> it.getId())
+			.collect(Collectors.toList());
+
+		assertThat(stationIds).containsExactlyElementsOf(expectedStationIds);
+	}
+
+	public static ExtractableResponse<Response> 지하철_노선에_지하철역_제외_요청(LineResponse line, StationResponse station) {
+		return RestAssured
+			.given().log().all()
+			.when().delete("/lines/{lineId}/sections?stationId={stationId}", line.getId(), station.getId())
+			.then().log().all()
+			.extract();
+	}
+
+	public static void 지하철_노선에_지하철역_제외됨(ExtractableResponse<Response> response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+	}
+
+	public static void 지하철_노선에_지하철역_제외_실패됨(ExtractableResponse<Response> response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
+	}
+
+}
